funding+rpcserver: move nodeHasTorAddrs to funding
We can simplify the funding manager's interface by moving the
nodeHasTorAddrs method to the funding package.
